Change global library-name
The library was defined two times on the window-context as window.Nexus
(from main.js), and as window.NexusUI (from webpack.config.js).

Now it is only defined once, as window.Nexus, if it is added to a page
with a script tag.

If Nexus is required/imported with webpack or similar it is not
available in the window-context anymore.